+++ app/src/main/java/com/sma2/sma2/ExerciseLogic/ExerciseSessionManager.java
@@ -2,9 +2,8 @@ package com.sma2.sma2.ExerciseLogic;
 
 import android.net.Uri;
 
-import com.sma2.sma2.ExReadText;
+import com.sma2.sma2.ExerciseFragments.ExReadText;
 import com.sma2.sma2.ExerciseFragments.ExAudioRec;
-import com.sma2.sma2.ExerciseFragments.ExOneFingerTapping;
 
 import java.util.ArrayList;
 import java.util.List;
