+++ app/src/main/res/layout/fragment_ex_read_text.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ExReadText">
+    tools:context=".ExerciseFragments.ExReadText">
 
     <ProgressBar
         android:id="@+id/volumeExRT"
